Increased unix time resolution on ESP-01 to milliseconds.
TimeSyncTester's output indicates the ESP-01's estimate of unix time is
about 20-60ms behind the system clock on the time server (the machine
running the Java program). This is actually a round-trip delay, so the
ESP-01 should be behind by less than that. ie, Both the time taken for a
datagram to go from the time server to the ESP-01, and the time taken
for a datagram to go from the ESP-01 to TimeSyncTester, contribute to
this delay.

The 20-60ms figure is similar to figures obtained from pinging the
ESP-01 on cmd.