Added filtering by location.
User can select location from <select> element. This gets included in
the query segment, of a url, in an HTTP request, sent by d3-request (a
layer over XHR).
DataServlet examines the query's 'loc' parameter, and uses its value to
filter for data from a particular location.

Ideally, should send query the database for the contents of the
Locations table, then populate the <select> element with that data,
instead of hardcoded in the html.